requirements.txt
pytesseract
pillow
discord.py
requests
mysql-connector